add note form: wip
